Move the trie approach to default.
