+++ lib/route-recognizer.js
@@ -1,362 +1,11 @@
-import map from './route-recognizer/dsl';
-import Normalizer from './route-recognizer/normalizer';
-
-var normalizePath = Normalizer.normalizePath;
-var normalizeSegment = Normalizer.normalizeSegment;
-
-var specials = [
-  '/', '.', '*', '+', '?', '|',
-  '(', ')', '[', ']', '{', '}', '\\'
-];
-
-var escapeRegex = new RegExp('(\\' + specials.join('|\\') + ')', 'g');
+/*
+TODO:
+- Support query params.
+- Support params for glob matching.
+*/
 
 function isArray(test) {
   return Object.prototype.toString.call(test) === "[object Array]";
-}
-
-// A Segment represents a segment in the original route description.
-// Each Segment type provides an `eachChar` and `regex` method.
-//
-// The `eachChar` method invokes the callback with one or more character
-// specifications. A character specification consumes one or more input
-// characters.
-//
-// The `regex` method returns a regex fragment for the segment. If the
-// segment is a dynamic of star segment, the regex fragment also includes
-// a capture.
-//
-// A character specification contains:
-//
-// * `validChars`: a String with a list of all valid characters, or
-// * `invalidChars`: a String with a list of all invalid characters
-// * `repeat`: true if the character specification can repeat
-
-function StaticSegment(string) { this.string = normalizeSegment(string); }
-StaticSegment.prototype = {
-  eachChar: function(currentState) {
-    var string = this.string, ch;
-
-    for (var i=0; i<string.length; i++) {
-      ch = string.charAt(i);
-      currentState = currentState.put({ invalidChars: undefined, repeat: false, validChars: ch });
-    }
-
-    return currentState;
-  },
-
-  regex: function() {
-    return this.string.replace(escapeRegex, '\\$1');
-  },
-
-  generate: function() {
-    return this.string;
-  }
-};
-
-function DynamicSegment(name) { this.name = normalizeSegment(name); }
-DynamicSegment.prototype = {
-  eachChar: function(currentState) {
-    return currentState.put({ invalidChars: "/", repeat: true, validChars: undefined });
-  },
-
-  regex: function() {
-    return "([^/]+)";
-  },
-
-  generate: function(params) {
-    if (RouteRecognizer.ENCODE_AND_DECODE_PATH_SEGMENTS) {
-      return encodeURIComponent(params[this.name]);
-    } else {
-      return params[this.name];
-    }
-  }
-};
-
-function StarSegment(name) { this.name = name; }
-StarSegment.prototype = {
-  eachChar: function(currentState) {
-    return currentState.put({ invalidChars: "", repeat: true, validChars: undefined });
-  },
-
-  regex: function() {
-    return "(.+)";
-  },
-
-  generate: function(params) {
-    return params[this.name];
-  }
-};
-
-function EpsilonSegment() {}
-EpsilonSegment.prototype = {
-  eachChar: function(currentState) {
-    return currentState;
-  },
-  regex: function() { return ""; },
-  generate: function() { return ""; }
-};
-
-// The `names` will be populated with {name, decode} objects for each
-// dynamic/star segment, where `name` is the parameter name for use during
-// recognition, and `decode` is whether the parameter value should be decoded
-// (true for dynamic segments, false for star segments).
-function parse(route, names, specificity) {
-  // normalize route as not starting with a "/". Recognition will
-  // also normalize.
-  if (route.charAt(0) === "/") { route = route.substr(1); }
-
-  var segments = route.split("/");
-  var results = new Array(segments.length);
-
-  // A routes has specificity determined by the order that its different segments
-  // appear in. This system mirrors how the magnitude of numbers written as strings
-  // works.
-  // Consider a number written as: "abc". An example would be "200". Any other number written
-  // "xyz" will be smaller than "abc" so long as `a > x`. For instance, "199" is smaller
-  // then "200", even though "y" and "z" (which are both 9) are larger than "0" (the value
-  // of (`b` and `c`). This is because the leading symbol, "2", is larger than the other
-  // leading symbol, "1".
-  // The rule is that symbols to the left carry more weight than symbols to the right
-  // when a number is written out as a string. In the above strings, the leading digit
-  // represents how many 100's are in the number, and it carries more weight than the middle
-  // number which represents how many 10's are in the number.
-  // This system of number magnitude works well for route specificity, too. A route written as
-  // `a/b/c` will be more specific than `x/y/z` as long as `a` is more specific than
-  // `x`, irrespective of the other parts.
-  // Because of this similarity, we assign each type of segment a number value written as a
-  // string. We can find the specificity of compound routes by concatenating these strings
-  // together, from left to right. After we have looped through all of the segments,
-  // we convert the string to a number.
-  specificity.val = '';
-
-  for (var i=0; i<segments.length; i++) {
-    var segment = segments[i], match;
-
-    if (match = segment.match(/^:([^\/]+)$/)) {
-      results[i] = new DynamicSegment(match[1]);
-      names.push({name: match[1], decode: true});
-      specificity.val += '3';
-    } else if (match = segment.match(/^\*([^\/]+)$/)) {
-      results[i] = new StarSegment(match[1]);
-      names.push({name: match[1], decode: false});
-      specificity.val += '1';
-    } else if(segment === "") {
-      results[i] = new EpsilonSegment();
-      specificity.val += '2';
-    } else {
-      results[i] = new StaticSegment(segment);
-      specificity.val += '4';
-    }
-  }
-
-  specificity.val = +specificity.val;
-
-  return results;
-}
-
-// A State has a character specification and (`charSpec`) and a list of possible
-// subsequent states (`nextStates`).
-//
-// If a State is an accepting state, it will also have several additional
-// properties:
-//
-// * `regex`: A regular expression that is used to extract parameters from paths
-//   that reached this accepting state.
-// * `handlers`: Information on how to convert the list of captures into calls
-//   to registered handlers with the specified parameters
-// * `types`: How many static, dynamic or star segments in this route. Used to
-//   decide which route to use if multiple registered routes match a path.
-//
-// Currently, State is implemented naively by looping over `nextStates` and
-// comparing a character specification against a character. A more efficient
-// implementation would use a hash of keys pointing at one or more next states.
-
-function State(charSpec, serialized) {
-  if (serialized) {
-    if (typeof serialized === "string") {
-      serialized = JSON.parse(serialized);
-    }
-
-    this.charSpec = serialized.charSpec;
-    this.nextStates = serialized.nextStates;
-    this.charSpecs = serialized.charSpecs;
-    this.regex = typeof serialized.regex !== "string" ? serialized.regex : new RegExp(serialized.regex);
-    this.handlers = serialized.handlers;
-    this.specificity = serialized.specificity;
-  } else {
-    this.charSpec = charSpec;
-    this.nextStates = [];
-    this.charSpecs = {};
-    this.regex = undefined;
-    this.handlers = undefined;
-    this.specificity = undefined;
-  }
-}
-
-State.prototype = {
-  get: function(charSpec) {
-    if (this.charSpecs[charSpec.validChars]) {
-      return this.charSpecs[charSpec.validChars];
-    }
-
-    var nextStates = this.nextStates;
-
-    for (var i=0; i<nextStates.length; i++) {
-      var child = nextStates[i];
-
-      var isEqual = child.charSpec.validChars === charSpec.validChars;
-      isEqual = isEqual && child.charSpec.invalidChars === charSpec.invalidChars;
-
-      if (isEqual) {
-        this.charSpecs[charSpec.validChars] = child;
-        return child;
-      }
-    }
-  },
-
-  put: function(charSpec) {
-    var state;
-
-    // If the character specification already exists in a child of the current
-    // state, just return that state.
-    if (state = this.get(charSpec)) { return state; }
-
-    // Make a new state for the character spec
-    state = new State(charSpec);
-
-    // Insert the new state as a child of the current state
-    this.nextStates.push(state);
-
-    // If this character specification repeats, insert the new state as a child
-    // of itself. Note that this will not trigger an infinite loop because each
-    // transition during recognition consumes a character.
-    if (charSpec.repeat) {
-      state.nextStates.push(state);
-    }
-
-    // Return the new state
-    return state;
-  },
-
-  // Find a list of child states matching the next character
-  match: function(ch) {
-    var nextStates = this.nextStates,
-        child, charSpec, chars;
-
-    var returned = [];
-
-    for (var i=0; i<nextStates.length; i++) {
-      if (nextStates[i] instanceof State) {
-        child = nextStates[i];
-      } else if (nextStates[i] === "__CIRCULAR__") {
-        child = this;
-      } else {
-        child = new State(undefined, nextStates[i]);
-      }
-
-      charSpec = child.charSpec;
-
-      if (typeof (chars = charSpec.validChars) !== 'undefined') {
-        if (chars.indexOf(ch) !== -1) { returned.push(child); }
-      } else if (typeof (chars = charSpec.invalidChars) !== 'undefined') {
-        if (chars.indexOf(ch) === -1) { returned.push(child); }
-      }
-    }
-
-    return returned;
-  },
-
-  serialize: function() {
-    return JSON.stringify(this, function(key, value) {
-      if (value === undefined) { return; }
-      if (key === "regex") { return value.source; }
-      if (key === "repeat" && !value) { return; }
-      if (Array.isArray(value) && value.length === 0) { return; }
-      if (typeof value === "object" && Object.keys(value).length === 0) { return; }
-
-      if (value && value.nextStates && value.nextStates.length && value === value.nextStates[0]) {
-        value.nextStates[0] = "__CIRCULAR__";
-        return value;
-      }
-
-      return value;
-    });
-  }
-};
-
-// Sort the routes by specificity
-function sortSolutions(states) {
-  return states.sort(function(a, b) {
-    return b.specificity.val - a.specificity.val;
-  });
-}
-
-function recognizeChar(states, ch) {
-  var nextStates = [];
-
-  for (var i=0, l=states.length; i<l; i++) {
-    var state;
-    if (states[i] instanceof State) {
-      state = states[i];
-    } else {
-      state = new State(undefined, states[i]);
-    }
-
-    nextStates = nextStates.concat(state.match(ch));
-  }
-
-  return nextStates;
-}
-
-var oCreate = Object.create || function(proto) {
-  function F() {}
-  F.prototype = proto;
-  return new F();
-};
-
-function RecognizeResults(queryParams) {
-  this.queryParams = queryParams || {};
-}
-RecognizeResults.prototype = oCreate({
-  splice: Array.prototype.splice,
-  slice:  Array.prototype.slice,
-  push:   Array.prototype.push,
-  length: 0,
-  queryParams: null
-});
-
-function findHandler(state, originalPath, queryParams) {
-  var handlers = state.handlers, regex = state.regex;
-  var captures = originalPath.match(regex), currentCapture = 1;
-  var result = new RecognizeResults(queryParams);
-
-  result.length = handlers.length;
-
-  for (var i=0; i<handlers.length; i++) {
-    var handler = handlers[i], names = handler.names || [], params = {};
-    var name, shouldDecode, capture;
-
-    for (var j=0; j<names.length; j++) {
-      name = names[j].name;
-      shouldDecode = names[j].decode;
-      capture = captures[currentCapture++];
-
-      if (RouteRecognizer.ENCODE_AND_DECODE_PATH_SEGMENTS) {
-        if (shouldDecode) {
-          params[name] = decodeURIComponent(capture);
-        } else {
-          params[name] = capture;
-        }
-      } else {
-        params[name] = capture;
-      }
-    }
-
-    result[i] = { handler: handler.handler, params: params, isDynamic: !!names.length };
-  }
-
-  return result;
 }
 
 function decodeQueryParamPart(part) {
@@ -369,77 +18,259 @@ function decodeQueryParamPart(part) {
   return result;
 }
 
-// The main interface
+function matcher(path, callback) {
+  var segments = path.replace(/^\//, '').split('/');
+  segments = segments.map(function(segment) {
+    return new SegmentTrieNode(segment);
+  });
 
-var RouteRecognizer = function(serialized) {
-  this.rootState = new State(undefined, serialized);
-  this.names = {};
+  var root = this;
+  var leaf = segments.reduce(function(previous, current) {
+    previous.append(current);
+    return current;
+  }, root);
+
+  if (callback) {
+    if (callback.length === 0) { throw new Error("You must have an argument in the function passed to `to`"); }
+    callback(matcher.bind(leaf));
+  }
+
+  return leaf;
+}
+
+function SegmentTrieNode(value) {
+  this.value = value;
+
+  switch (this.value.charAt(0)) {
+    case ':': this.type = 'param'; break;
+    case '*': this.type = 'glob'; break;
+    default: this.type = 'static'; break;
+  }
+
+  this.handler = undefined;
+  this.childNodes = [];
+}
+
+SegmentTrieNode.prototype = {
+  append: function(trieNode) {
+    this.childNodes.push(trieNode);
+    trieNode.parentNode = this;
+    return this;
+  },
+
+  compact: function() {
+    if (this.childNodes.length === 0) { return; }
+
+    // Depth-first compaction.
+    this.childNodes.forEach(function(trieNode) {
+      trieNode.compact();
+    });
+
+    // Collapse sibling nodes.
+    this.childNodes = this.childNodes.filter(function(trieNode, index, siblingNodes) {
+      var segmentSeen = false;
+
+      // Scan only segments before this one to see if we've already got a match.
+      for (var i = 0; i < index && segmentSeen === false; i++) {
+        segmentSeen = (
+          siblingNodes[i].value === trieNode.value &&
+          siblingNodes[i].handler === trieNode.handler
+        );
+      }
+
+      if (segmentSeen) {
+        var targetNode = siblingNodes[i-1];
+
+        // Reset the parentNode for each trieNode to the targetNode.
+        trieNode.childNodes.forEach(function(trieNode) {
+          trieNode.parentNode = targetNode;
+        });
+
+        // Concat the childNodes of the active trieNode with the targetNode.
+        targetNode.childNodes = targetNode.childNodes.concat(trieNode.childNodes);
+
+        // Then re-compact the joined trie.
+        targetNode.compact();
+      }
+
+      return !segmentSeen;
+    });
+
+    this.childNodes.sort(function(a, b) {
+      var ascore, bscore;
+      switch (a.type) {
+        case "static": ascore = 0; break;
+        case "param": ascore = 1; break;
+        case "glob": ascore = 2; break;
+      }
+      switch (b.type) {
+        case "static": bscore = 0; break;
+        case "param": bscore = 1; break;
+        case "glob": bscore = 2; break;
+      }
+
+      return ascore > bscore;
+    });
+  },
+
+  to: function(handler, callback) {
+    this.handler = handler;
+    if (callback) {
+      if (callback.length === 0) { throw new Error("You must have an argument in the function passed to `to`"); }
+      callback(matcher.bind(this));
+    }
+    return this;
+  },
+
+  /**
+    Our goal is to try and match based upon the node type. For non-globbing
+    routes we can simply pop a segment off of the path and continue, eliminating
+    entire branches as we go. Average number of comparisons is:
+      `Number of Trie Nodes / Average Depth / 2`
+
+    If we reach a globbing route we have to change strategy and traverse to all
+    descendent leaf nodes until we find a match. As we traverse we build up a
+    regular expression that would match beginning with that globbing route. We
+    leverage the regular expression to handle the mechanics of greedy pattern
+    matching with back-tracing. The average number of comparisons beyond a
+    globbing route:
+      `Number of Trie Nodes / 2`
+
+    This could be optimized further to do O(1) matching for non-globbing
+    segments but that is overkill for this use case.
+  */
+  walk: function(path, handlers, params, regexPieces) {
+    var isLeafNode = (this.childNodes.length === 0);
+    var nodeMatches = false;
+    var nextNode = this;
+    var consumed = false;
+    var segmentIndex = 0;
+
+    // Identify the node type so we know how to match it.
+    switch (this.type) {
+      case "static":
+        if (regexPieces) {
+          // If we're descended from a globbing route.
+          regexPieces.push(this.value);
+        } else {
+          // Matches if the path to recognize is identical to the node value.
+          segmentIndex = this.value.length + 1;
+          nodeMatches = path.indexOf(this.value) === 0;
+          if (nodeMatches) { path = path.substring(segmentIndex); }
+        }
+      break;
+      case "param":
+        if (regexPieces) {
+          // If we're descended from a globbing route.
+          regexPieces.push('([^/]+)');
+        } else {
+          // Valid for '/' to not appear, or appear anywhere but the 0th index.
+          // 0 length or 0th index would result in an non-matching empty param.
+          segmentIndex = path.indexOf('/');
+          if (segmentIndex === -1) { segmentIndex = path.length; }
+
+          nodeMatches = (path.length > 0 && segmentIndex > 0);
+          if (nodeMatches) {
+            params[this.value.substring(1)] = path.substr(0, segmentIndex);
+            path = path.substring(segmentIndex + 1);
+          }
+        }
+      break;
+      case "glob":
+        // We can no longer do prefix matching. Prepare to traverse leaves.
+
+        // It's possible to have multiple globbing routes in a single path.
+        // So maybe we already have a `regexPieces` array.
+        if (!regexPieces) { regexPieces = []; }
+        params[this.value.substring(1)] = regexPieces.length;
+
+        if (isLeafNode) {
+          // If a glob is the leaf node we don't match a trailing slash.
+          regexPieces.push('(.+)(?:/?)');
+        } else {
+          // Consume the segment. `regexPieces.join` adds the '/'.
+          regexPieces.push('(.+)');
+        }
+      break;
+    }
+
+    // Short-circuit for nodes that can't possibly match.
+    if (!nodeMatches && !regexPieces) { return false; }
+
+    if (this.handler) {
+      handlers.push({
+        handler: this.handler,
+        params: params,
+        isDynamic: (this.type !== "static")
+      });
+    }
+
+    var nextParams = this.handler ? {} : params;
+
+    // Depth-first traversal of childNodes. No-op for leaf nodes.
+    for (var i = 0; i < this.childNodes.length; i++) {
+      nextNode = this.childNodes[i].walk(path, handlers, nextParams, regexPieces);
+
+      // Stop traversing once we have a match since we're sorted by specificity.
+      if (!!nextNode) { break; }
+    }
+
+    // If we're at a leaf node find out if we've consumed the entire path.
+    if (isLeafNode) {
+      if (regexPieces) {
+        consumed = new RegExp('^'+regexPieces.join('/')+'$').test(path);
+  
+        // We pushed a segment onto the regexPieces, but this wasn't a match.
+        // Pop it back off for the next go-round.
+        if (!consumed) { regexPieces.pop(); }
+      } else {
+        consumed = (path.length === 0);
+      }
+    }
+
+    // `consumed` is false unless set above.
+    if (isLeafNode && !consumed) {
+      return false;
+    } else {
+      return nextNode;
+    }
+  }
 };
 
+function RouteRecognizer(serialized) {
+  this.names = {};
+  this.rootState = new SegmentTrieNode('');
+}
 
 RouteRecognizer.prototype = {
   add: function(routes, options) {
-    var currentState = this.rootState, regex = "^",
-        specificity = {},
-        handlers = new Array(routes.length), allSegments = [], name;
-
-    var isEmpty = true;
-
-    for (var i=0; i<routes.length; i++) {
-      var route = routes[i], names = [];
-
-      var segments = parse(route.path, names, specificity);
-
-      allSegments = allSegments.concat(segments);
-
-      for (var j=0; j<segments.length; j++) {
-        var segment = segments[j];
-
-        if (segment instanceof EpsilonSegment) { continue; }
-
-        isEmpty = false;
-
-        // Add a "/" for the new segment
-        currentState = currentState.put({ invalidChars: undefined, repeat: false, validChars: "/" });
-        regex += "/";
-
-        // Add a representation of the segment to the NFA and regex
-        currentState = segment.eachChar(currentState);
-        regex += segment.regex();
-      }
-      var handler = { handler: route.handler, names: names };
-      handlers[i] = handler;
+    options = options || {};
+    var route;
+    var leaf = this.rootState;
+    while (routes.length) {
+      route = routes.shift();
+      leaf = matcher.bind(leaf)(route.path);
+      leaf.to(route.handler);
     }
-
-    if (isEmpty) {
-      currentState = currentState.put({ invalidChars: undefined, repeat: false, validChars: "/" });
-      regex += "/";
-    }
-
-    currentState.handlers = handlers;
-    currentState.regex = new RegExp(regex + "$");
-    currentState.specificity = specificity;
-
-    if (name = options && options.as) {
-      this.names[name] = {
-        segments: allSegments,
-        handlers: handlers
-      };
-    }
+    leaf.name = options.as;
+    this.names[options.as] = leaf;
+    this.rootState.compact();
   },
 
   handlersFor: function(name) {
-    var route = this.names[name];
+    var trieNode = this.names[name];
 
-    if (!route) { throw new Error("There is no route named " + name); }
+    if (!trieNode) { throw new Error("There is no route named " + name); }
 
-    var result = new Array(route.handlers.length);
+    var handlers = [];
 
-    for (var i=0; i<route.handlers.length; i++) {
-      result[i] = route.handlers[i];
-    }
+    do {
+      if (trieNode.handler) {
+        handlers.push(trieNode.handler);
+      }
+    } while (trieNode = trieNode.parentNode);
 
-    return result;
+    return handlers;
   },
 
   hasRoute: function(name) {
@@ -447,30 +278,9 @@ RouteRecognizer.prototype = {
   },
 
   generate: function(name, params) {
-    var route = this.names[name], output = "";
-    if (!route) { throw new Error("There is no route named " + name); }
-
-    var segments = route.segments;
-
-    for (var i=0; i<segments.length; i++) {
-      var segment = segments[i];
-
-      if (segment instanceof EpsilonSegment) { continue; }
-
-      output += "/";
-      output += segment.generate(params);
-    }
-
-    if (output.charAt(0) !== '/') { output = '/' + output; }
-
-    if (params && params.queryParams) {
-      output += this.generateQueryString(params.queryParams, route.handlers);
-    }
-
-    return output;
   },
 
-  generateQueryString: function(params, handlers) {
+  generateQueryString: function(params) {
     var pairs = [];
     var keys = [];
     for(var key in params) {
@@ -500,6 +310,11 @@ RouteRecognizer.prototype = {
     if (pairs.length === 0) { return ''; }
 
     return "?" + pairs.join("&");
+  },
+
+  map: function(callback, addRouteCallback) {
+    callback(matcher.bind(this.rootState));
+    this.rootState.compact();
   },
 
   parseQueryString: function(queryString) {
@@ -533,77 +348,15 @@ RouteRecognizer.prototype = {
   },
 
   recognize: function(path) {
-    var states = [ this.rootState ],
-        pathLen, i, l, queryStart, queryParams = {},
-        hashStart,
-        isSlashDropped = false;
-
-    hashStart = path.indexOf('#');
-    if (hashStart !== -1) {
-      path = path.substr(0, hashStart);
-    }
-
-    queryStart = path.indexOf('?');
-    if (queryStart !== -1) {
-      var queryString = path.substr(queryStart + 1, path.length);
-      path = path.substr(0, queryStart);
-      queryParams = this.parseQueryString(queryString);
-    }
-
-    if (path.charAt(0) !== "/") { path = "/" + path; }
-    var originalPath = path;
-
-    if (RouteRecognizer.ENCODE_AND_DECODE_PATH_SEGMENTS) {
-      path = normalizePath(path);
-    } else {
-      path = decodeURI(path);
-      originalPath = decodeURI(originalPath);
-    }
-
-    pathLen = path.length;
-    if (pathLen > 1 && path.charAt(pathLen - 1) === "/") {
-      path = path.substr(0, pathLen - 1);
-      originalPath = originalPath.substr(0, pathLen - 1);
-      isSlashDropped = true;
-    }
-
-    for (i=0; i<path.length; i++) {
-      states = recognizeChar(states, path.charAt(i));
-      if (!states.length) { break; }
-    }
-
-    var solutions = [];
-    for (i=0; i<states.length; i++) {
-      if (states[i].handlers) { solutions.push(states[i]); }
-    }
-
-    states = sortSolutions(solutions);
-
-    var state = solutions[0];
-
-    if (state && state.handlers) {
-      // if a trailing slash was dropped and a star segment is the last segment
-      // specified, put the trailing slash back
-      if (isSlashDropped && state.regex.source.slice(-5) === "(.+)$") {
-         originalPath = originalPath + "/";
-       }
-      return findHandler(state, originalPath, queryParams);
-    }
+    var handlers = [];
+    var trieNode = this.rootState.walk(path, handlers, {});
+    return handlers;
   },
 
   toJSON: function() {
-    return this.rootState.serialize();
   }
 };
 
-RouteRecognizer.prototype.map = map;
-
 RouteRecognizer.VERSION = 'VERSION_STRING_PLACEHOLDER';
-
-// Set to false to opt-out of encoding and decoding path segments.
-// See https://github.com/tildeio/route-recognizer/pull/55
-RouteRecognizer.ENCODE_AND_DECODE_PATH_SEGMENTS = true;
-
-RouteRecognizer.Normalizer = Normalizer;
 
 export default RouteRecognizer;
