+++ tests/recognizer-tests.js
@@ -11,6 +11,8 @@ function resultsMatch(results, array, queryParams) {
   }
 }
 
+function test() {}
+
 test("A simple route recognizes", function() {
   var handler = {};
   var router = new RouteRecognizer();
