+++ tests/router-tests.js
@@ -13,12 +13,7 @@ function resultsMatch(results, array, queryParams) {
 
 module("The match DSL", {
   setup: function() {
-    router.map = function() {
-      var original = new RouteRecognizer();
-      original.map.apply(original, arguments);
-      var serialized = original.toJSON();
-      router = new RouteRecognizer(serialized);
-    };
+    router = new RouteRecognizer();
   }
 });
